demo_api/requirements.txt
Django==1.11.7
djangorestframework==3.7.3
docutils==0.14
